Rebuilt UI, performance, library functionality.
Most of the UI components were rebuilt making the app look more native.
Added a custom icon for the app.
Added a loader.
The home background video is paused when not used to improve app's performance.
Added visual effects for when the app is either blurred or focused.
An custom background image can be set when adding new games.
- A preview of this image was added to the game editor.
The games' data folder can be openned through each game menu.
Library functionality:
- Search games.
- Add games from library.